requirements-deploy.txt
altair==5.1.2
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.1
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
plotly==5.18.0
protobuf==4.25.0
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.31.0
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
streamlit==1.28.2
streamlit-option-menu==0.3.6
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
xgboost==2.0.2
zipp==3.17.0
Flask==3.0.0
Flask-CORS==4.0.0
tensorflow==2.15.0
keras==3.0.0
h5py==3.10.0
Werkzeug==3.0.1
